split specs2 traits into separate files
